+++ mobidam-sst-management-frontend/frontend/src/api/SchnittstelleService.ts
@@ -23,6 +23,9 @@
 
 import FetchUtils from "@/api/FetchUtils";
 import Schnittstelle from "@/types/Schnittstelle";
+import { useSnackbarStore } from "@/stores/snackbar";
+import { Levels } from "@/api/error";
+import SchnittstelleRequest from "@/types/SchnittstelleRequest";
 
 export default class SchnittstelleService {
     private static base: string | undefined = import.meta.env
@@ -41,5 +44,29 @@ export default class SchnittstelleService {
             }
             return response.json();
         });
+    }
+
+    public static create(
+        instance: SchnittstelleRequest
+    ): Promise<Schnittstelle> {
+        return fetch(
+            `${this.base}/api/schnittstelle`,
+            FetchUtils.getPOSTConfig(instance)
+        )
+            .then((response) => {
+                useSnackbarStore().showMessage({
+                    message: "Speichern erfolgreich.",
+                    level: Levels.SUCCESS,
+                });
+                FetchUtils.defaultResponseHandler(response);
+                return response.json();
+            })
+            .catch((err) => {
+                useSnackbarStore().showMessage({
+                    message: "Speichern der Schnittstelle fehlgeschlagen.",
+                    level: Levels.ERROR,
+                });
+                FetchUtils.defaultResponseHandler(err);
+            });
     }
 }