+++ mobidam-sst-management-frontend/frontend/src/views/MainView.vue
@@ -32,7 +32,16 @@
             </v-col>
         </v-row>
         <v-row>
-            <h2>Schnittstellen</h2>
+            <h2>
+                Schnittstellen &nbsp;
+                <v-btn
+                    small
+                    outlined
+                    @click="showAddSchnittstelleDialog = true"
+                >
+                    <v-icon>mdi-plus</v-icon>
+                </v-btn>
+            </h2>
         </v-row>
         <br />
         <v-list>
@@ -119,6 +128,10 @@
                 </v-tooltip>
             </v-list-item>
         </v-list>
+        <add-schnittstelle-dialog
+            :show-dialog.sync="showAddSchnittstelleDialog"
+            @schnittstelle-saved="refreshList"
+        ></add-schnittstelle-dialog>
     </v-container>
 </template>
 
@@ -130,8 +143,10 @@ import SchnittstelleService from "@/api/SchnittstelleService";
 import Datentransfer from "@/types/Datentransfer";
 import DatentransferService from "@/api/DatentransferService";
 import SchnittstelleWithDatentransfer from "@/types/SchnittstelleWithDatentransfer";
+import AddSchnittstelleDialog from "@/components/AddSchnittstelleDialog.vue";
 
 const snackbarStore = useSnackbarStore();
+const showAddSchnittstelleDialog = ref(false);
 const schnittstellen = ref<SchnittstelleWithDatentransfer[]>([]);
 const sortedSchnittstellen = computed(() => {
     let sorted = ref<SchnittstelleWithDatentransfer[]>([]);
@@ -191,6 +206,11 @@ function getDatentransferZeitstempel(
     if (schnittstelle.datentransfer)
         return schnittstelle.datentransfer.zeitstempel;
     return "";
+}
+
+function refreshList(): void {
+    schnittstellen.value = [];
+    getSchnittstellen();
 }
 </script>